Cleanup folder structure
