+++ .gitignore
@@ -2,3 +2,4 @@
 cuHacking/GoogleService-Info.plist
 .map-box
 ./Pods/
+xcuserdata/
