+++ cuHacking/UI/TitleImageView.swift
@@ -33,7 +33,6 @@ class TitleImageView: UICollectionReusableView {
     override init(frame: CGRect) {
         super.init(frame: frame)
         setup()
-        
     }
 
     required init?(coder: NSCoder) {
